+++ src/main/java/br/com/thiagoRDS/api_authors/modules/authors/usecases/AuthenticateAuthor.java
@@ -9,7 +9,7 @@ import br.com.thiagoRDS.api_authors.modules.authors.entities.Author;
 import br.com.thiagoRDS.api_authors.modules.authors.exceptions.InvalidCredentialsException;
 import br.com.thiagoRDS.api_authors.modules.authors.repositories.AuthorsRepository;
 import br.com.thiagoRDS.api_authors.providers.JwtProvider.JwtProvider;
-import br.com.thiagoRDS.api_authors.providers.dtos.SignResponseDTO;
+import br.com.thiagoRDS.api_authors.providers.JwtProvider.dtos.SignResponseDTO;
 
 @Service
 public class AuthenticateAuthor {
