+++ src/main/java/br/com/thiagoRDS/api_authors/providers/JwtProvider/JwtProvider.java
@@ -2,7 +2,7 @@ package br.com.thiagoRDS.api_authors.providers.JwtProvider;
 
 import com.auth0.jwt.interfaces.DecodedJWT;
 
-import br.com.thiagoRDS.api_authors.providers.dtos.SignResponseDTO;
+import br.com.thiagoRDS.api_authors.providers.JwtProvider.dtos.SignResponseDTO;
 
 public abstract class JwtProvider {
   public abstract SignResponseDTO sign(String authorId);
