+++ src/main/java/br/com/thiagoRDS/api_authors/providers/JwtProvider/dtos/SignResponseDTO.java
@@ -1,0 +1,6 @@
+package br.com.thiagoRDS.api_authors.providers.JwtProvider.dtos;
+
+public record SignResponseDTO(
+                String token,
+                Long expiresIn) {
+}
