--- src/main/java/br/com/thiagoRDS/api_authors/providers/dtos/SignResponseDTO.java
@@ -1,6 +1,0 @@
-package br.com.thiagoRDS.api_authors.providers.dtos;
-
-public record SignResponseDTO(
-        String token,
-        Long expiresIn) {
-}
