+++ src/test/java/br/com/thiagoRDS/api_authors/modules/authors/controllers/DeleteAuthorControllerTest.java
@@ -32,7 +32,7 @@ import br.com.thiagoRDS.api_authors.modules.utils.MakeAddress;
 import br.com.thiagoRDS.api_authors.modules.utils.MakeAuthor;
 import br.com.thiagoRDS.api_authors.modules.utils.MakePost;
 import br.com.thiagoRDS.api_authors.providers.JwtProvider.JwtProvider;
-import br.com.thiagoRDS.api_authors.providers.dtos.SignResponseDTO;
+import br.com.thiagoRDS.api_authors.providers.JwtProvider.dtos.SignResponseDTO;
 
 @Transactional
 @ActiveProfiles("test")
