+++ src/test/java/br/com/thiagoRDS/api_authors/modules/authors/controllers/UploadAvatarControllerTest.java
@@ -27,7 +27,7 @@ import br.com.thiagoRDS.api_authors.modules.authors.repositories.AuthorsReposito
 import br.com.thiagoRDS.api_authors.modules.utils.MakeAuthor;
 import br.com.thiagoRDS.api_authors.providers.MinioProvider;
 import br.com.thiagoRDS.api_authors.providers.JwtProvider.JwtProvider;
-import br.com.thiagoRDS.api_authors.providers.dtos.SignResponseDTO;
+import br.com.thiagoRDS.api_authors.providers.JwtProvider.dtos.SignResponseDTO;
 
 @Transactional
 @ActiveProfiles("test")
